Removed closure with new GamePlay Trait
Previously, the game play logic was held in a closure and the states in
local variables. That was an ad-hoc solution for a POC.

Now the GamePlay Trait provides a more idiomatic layout.

In 'olcConsoleGameEngine' user needs to create a sub class and override
few methods where he/she inserts the game play and drawing logic.
The state also resides in the same class.

In Rust, it made more sense to keep the Game Engine logic separate and
instead of 'composing', we pass a GamePlay instance. This instance
abstracts the drawing, game play logic and states from the Engine.